UTET-1 Add self-stretching background sprite
Fits to the camera with each frame (hopefully not performance-harming).
Playing around a little bit with unit tests. I tried writing something
for this script but it was proving a pain, since it references a camera,
Screen, SpriteRenderer, etc. Perhaps there's a nice way to push through
it, but I'll find an easier script to test in a later change, for a
start.

I'm unsure at this stage whether a self-stretching sprite is the go-to
for a background, or if it should be a UI element. The latter seems
straightforward, but that'd mean rendering sprites (bricks) on top of UI
elements, which is also very odd. So going with the sprite approach, for
now.