UTET-3 Implement L block movement, snapping, and spawning loop
Still no automatic falling or rotation, but after a bunch of back and
forth on how to best model the game's backend, I've settled on a
practical mix of Blocks knowing its coordinates and having those
updated, along with the PlayAreaController having its own backing
Dictionary for what's where. I would have liked to eliminate one of them
cause I didn't want to have to have to update this state in two places,
but I unfortunately don't see a better way since we need to know where a
block is in order move it (naturally).