UTET-12 Implement L block rotation (with wall kick translations)
This was one of the trickier ones! Wall kick translations refers to
translations we apply to the rotated block and test when the original
translation is invalid (due to overlapping with something or being out
of bounds).

These wall kick translations require a hard-coded matrix of coordinates
to try for different pieces (Line block is really the only unique one)
and depending on the translation. I used the following link for those:

https://tetris.fandom.com/wiki/SRS

We currently only have clockwise rotations working (since that's how I
usually play exclusively - I am not a good Tetris player...). I've
raised UTET-32 to implement anticlockwise rotations however, and the
code should not make that too hard.

I've also realized that I've made a mistake of implementing Coordinate -
I forgot Vector2Int existed and would have been a nice thing to just use
instead, but oh well. We're learning!