UTET-30 Implement gravity
Implement the mechanism for automatically moving down the block on an
interval. We introduce some Action events to facilitate handling this.
We reset the timer if the player does a deliberate down-move themselves,
just so that we don't surprise them with a "double" move down if they
happen to move down, coinciding with the timer.

We also properly detect when the downwards move is due to a wall kick
and not the player deliberately moving downwards. This requires some
additional information being applied to BlockTransformation.

The gravity amount is currently constant. We can implement something
more advanced i.e. scaling with time/score when we later add difficulty
to the game (UTET-7).