UTET-11 Implement instant place with spacebar
By pressing space, we will now place the block instantly. This is not a
complete solution - we're not properly handling what happens as the
blocks reach the spawn. Right now, we'll just throw an exception if we
try to place, otherwise we just keep spawning blocks and immediately
placing them on top of each other. This should be addressed in UTET-6.

We take a (perhaps naive) approach of just iterating from the bottom to
the top of the play area, placing at the first valid spot.