UTET-34 Spawn block higher into "pillow room" as blocks build up
Usually, the spawned block will be placed within the 0 - 1 row index
range. However, as blocks are built up to block those two rows, we want
to place the spawning block higher so as to not intersect with these
already-placed pieces.

We thus introduce this "pillow room" by adding two additional cell rows:
-1 and -2. These are represented by a transparent red.