UTET-35 Bug fix: Change instant place approach to bottom-down checking
We implement instant placement in UTET-11, but we took a bottom-up
approach. In hindsight, this obviously doesn't work, because we'll end
up teleporting into block-shaped holes that may be buried underneath
other blocks, rather than placing on top of them, as expected.

This commit quickly changes that approach to checking positions
*downwards*. When we find a row that fail, we place on the previous one.
This seems to work very well!