UTET-36 Center block spawning
Blocks currently always spawn on the left side of the play area. Let's
add an offset for each block to center them. Since Tetris has 10
horizontal cells, centering most blocks that are 3 cells wide isn't
exactly possible, and we prefer the "left-center" spawn point.

We need to add this additional centering offset to the Block's
constructor so that it can update its internal pivot state to match.