UTET-9 Add ghost block preview
It can currently be a bit difficult to instant place blocks since you
have to eye ball where it will fall. As a quality of life, let's add a
ghost preview of where the block will be placed if the user decides to
hit enter.

The approach we use is pretty simple. We actually re-use the instant
place calculation and, instead of placing the block there, we add a new
"ghost block piece" to the game cell at the output coordinates. The
GameCell will then take care of rendering itself slightly differently to
form a "ghost preview".

We do need to add some more methods and state to PlayAreaController. I
should think about splitting that thing up since it's getting quite
large!