UTET-6 Stop spawning blocks on game over
We now detect when a block is placed within the "pillow room",
signifying a game over. I'm trying out a global "GameState" system that
every MonoBehavior can reference via getting from the GameController,
which will probably be persistent through scenes. We'll find out whether
this is a nice pattern or not!