UTET-6 Implement play again button on game over
This button appears when the player loses. Clicking it hides it again,
and restarts the game. It triggers a GameStarted event via our new
GameState object that classes such as PlayAreaController listen to and
then clean up their state, ready for a new game to begin.

Disabling the button was surprisingly unintuitive. The go-to way appears
to be to set the entire GameObject as inactive, but that then disables
the Button script as well, which I want to have responsible for unhiding
itself when the game is over. There are some suggestions online to put
that code elsewhere, but that's not very clean I think, from an OOP
perspective.

The solution, suggested by jackalkane in the below link, is to have a
GameObject elsewhere in the scene. When I want to hide the button, I set
the button's transform parent to this GameObject instead of the Canvas,
and vice versa to re-enable it.

Seems to work well, but seems a bit hacky at the same time. There appear
to be a multitude of ways to do this - this seems like the simplest. I
wonder if there's a supposedly "correct" way to do it, though.

https://forum.unity.com/threads/how-to-disable-hide-ui-elements-without-disabling-them.282100/