UTET-7 Add increasing difficulty
Difficulty right now only controls how the speed at which blocks fall
scales over time. This is probably the only way to exact difficulty in
Tetris that I'll be implementing. The difficulty is controlled in two
ways: by a normalized [0, 1] "difficulty" factor and a [0, 0.08] "speed
rampup factor".

The former generally controls how fast pieces fall. If you imagine a
plot of time to drop time, difficulty controls the Y translation of the
graph.

The rampup factor controls how quickly the difficulty scales up. Higher
values scales more aggressively with time.

I intend for the rampup factor to be a developer-set constant to what
"feels" right. The player will later be able to adjust their gamea
difficulty, but it will only affect the first "difficulty" factor, not
the rampup factor.

That's the intention at this time, anyway.