UTET-37 Implement block preview panel
We add a panel on the right which displays the three next blocks that
will be played. This change contains various refactors relating to
making this work. For example, to actually display those pieces, we need
to generate additional GameCells on which to render them, separately
from the play area. We also need to alter some of the centering logic
because the dimensions are different between the play area and the block
preview panel.

Also update the gameplay GIF with the latest features in the README.md.