UTET-10 Implement block stashing
Using the shift key, the player can now swap their current block in play
into a stash on the left. If there's no block there, we simply spawn the
next one from the spawn preview panel on the right. This will only
happen for the first swap. Later swaps will actually have a block there,
in which case we do just that - put the stashed block in play, and put
the block that's currently already in play into the stash.

Players can only stash once per block spawn i.e. cannot keep swapping
endlessly. The ability to stash resets whenever a block is placed.

Swapping reset the current block height, so can be used as a bit of
"second breath" in a placement.

Stashing a block resets its rotation. Main reason for this is so that
the block stash area doesn't need to be sized to fit all possible blocks
and their rotations. Also makes it easier to control the actual
placement when swapping back into the play area.