UTET-33 Implement "bag of blocks" spawning algorithm
Rather than just having a weighed block spawning algorithm, it seems
like the most common algorithm according to the Tetris guidelines is a
"bag of blocks" approach, which I quite like as well for its simplicity
and positive impact on gameplay. TLDR: Put all 7 blocks in a bag, pull
out blocks randomly one at a time, when bag is empty, re-fill it again.