UTET-38 Add IOriginProvider
When implementing the game over screen, I want to shift the play area to
the left. In order to do this though, I need to modify the origin
coordinate for all the Coordinate objects. Therefore, rather than them
holding a static float for this origin, give them a provider, which will
change and thus shift all the cells over as well.