UTET-38 Add game over text and title
The text currently has some placeholder logic. Once highscores are
implemented (UTET-21), we will change this to comment on whether or not
you set a new high score. We display it after a short delay (intended
for when the UI shifts have completed).

There's some blank spaces at the moment - UTET-45 will add a UI element
to display the high scores.