UTET-47 Fix spawning bug that would clear some cells
At some point I introduced a bug when shifting the spawning piece into
the center, which could clear existing cells if there were any where
it's initially positioned after being spawned (before centering).

We fix that now. The PlayAreaController class is truly monstrous but I
think I'm pretty much done touching that class now that it's pretty
feature complete and bug-less. It's certainly a lesson for the future,
though.

The only potential is maybe wanting to add some fancy line completion
animation or something like it, but we'll see.