UTET-14 Implement pause menu with working resume, restart, and quit
The pause menu properly pauses time. The restart properly resets the
game for another go (I did notice a bug which I've raised UTET-48 for,
but it was pre-existing), and the quit button returns the user to the
main menu. This last bit means we now have a complete cyclic graph of
the player being able to navigate from the main menu, into the game,
back to the menu!