UTET-20 Implement difficulty slider
When the player navigates from the main menu to the gameplay, they go
through a menu with a difficulty slider that previously did nothing. We
now use that slider to set a difficulty parameter and remember it into
the next scene using our now-DDOL (DontDeleteOnLoad) GameState object.
This involved a few other code changes to facilitate.

For example, each scene has its own GameController with a GameState so
that I can press play in the editor from any of them. However, we only
want one at a time so if the player navigates from the main menu to the
gameplay scene, the latter's GameController will destroy itself after
detecting one already exists.

I was previously depending on a few in-editor wirings for button which
would directly talk to the GameController, so those broke. However, I'm
come up with a pattern which includes each scene still having their own
GameControllerProxy that will just forward the calls dynamically to the
true GameState, but can still be used for in-editor wirings.

This all seems to work quite well, working through many main menu <->
gameplay scene navigations!