UTET-49 Add quit to desktop button in pause menu & resize gameplay UI
First part is self-explanatory (there is now a quit to main menu *and* a
quit to desktop button in the pause menu), but for the second part, we
now scale the UI correctly with changing window sizes. The constraints
before were incorrect so they'd get messed up on certain resolutions,
but things should now scale correctly, thanks to changing some settings
in the Canvas object.