UTET-21 Store highscores transiently
We don't currently do anything with it, just debug log the saves we're
making. Since we're saving it though, the next step will be to actually
display these highscores on the game over screen. We will eventually
make these persisted as well, writing to the file system.

For now we generate a player name, but the idea is also to let the
player enter this themselves.