UTET-22 Persist high scores
They're now no longer transient i.e. we store high scores in a JSON file
on the computer's disk space. Not very secure - the JSON is plain text
and can be edited, but it's a single player game so I'm not too
concerned. Plus, I kinda have an idea for what I'd do to counter it e.g.
encrypt and/or store in binary instead.

I tried to come up with some sort of nice design for this all being
handled, with some abstractions, but I'm not convinced of it. Not sure
of the best way for save data information to be accessed and stored by
the rest of my game. I took the approach of having a "SaveData" object
that clients e.g. HighScoreManager can read from and update however they
like. They can then request that the SaveManager actually save after the
changes have been made.

I'm also leaking that we're using JSON, somewhat. Not sure of the best
way to counter that, needs some more thought.

I'm using DTOs when actually writing to disk to try and allow changing
persistence logic without effecting the rest of the game code. I didn't
do that for SaveData though, which maybe I should have done.

I think this is a reasonable initial implementation - I'll try to learn
from it for future projects.