UTET-18 UTET-51 Add music and volume slider to options menu
We're simply playing a tetris theme on loop - no fancy syncing to
gameplay or anything else that's happening.

Also begin implementing the options menu. We have the volume slider
working, and a no-op fullscreen toggle showing as well (just as a
placeholder for an additional setting, to help with layout).