UTET-54 Filter resolutions of different refresh rates
Turns out the Resolution object is really just a simple struct
containing the weight, height, and crucially, the refresh rate. Not sure
why that's part of it, but it explains the "duplicate" resolutions I saw
when originally implementing this. Let's filter to only get resolutions
equivalent to the current one.