UTET-29 Add button for deleting all high scores
So if the player ever wants to clear their high scores list, they now
have an in-game ability to do that (i.e. no need to mess with the
save files).

This button is somewhat dangerous, so I've raised a lo-prio ticket
UTET-57 to add a 'are you sure?' dialog pop-up.